Refactor import statements in footer components
